Ensure creation of 'debug-sites' directory before access.
Even though the 'debug-sites' directory is present in a checked out version of this
software it may be deleted by a user in order to clean up the debug logs.
Therefore the directory is removed from version-control and created by the software
if not already present before the first access.